docker_setup/model_server/requirements.txt
torch==2.1.0
transformers==4.36.2
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
PyYAML==6.0.1 